@@ -17,7 +17,6 @@ limitations under the License.
 package kv
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -31,7 +30,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-//TestPutBranch verifies putBranch function.
+// TestPutBranch verifies putBranch function.
 func TestPutBranch(t *testing.T) {
 	var repo repository.Repo
 	_, path, _, _ := runtime.Caller(0)
@@ -45,14 +44,18 @@ func TestPutBranch(t *testing.T) {
 
 	handler.putBranch(repo, repo.Branch()) //nolint:errcheck
 	handler.Commit()                       //nolint:errcheck
-
-	err := filepath.Walk(repository.WorkDir(repo), func(path string, f os.FileInfo, err error) error { //nolint:staticcheck
+	var wd = repository.WorkDir(repo)
+	err := filepath.Walk(wd, func(path string, f os.FileInfo, err error) error { //nolint:staticcheck
 		// Skip the .git directory
 		if f.IsDir() && f.Name() == ".git" {
 			return filepath.SkipDir
 		}
+		// Just push text files
+		if !f.IsDir() && filepath.Ext(f.Name()) != ".go" {
+			return nil
+		}
 
-		// Do not push directories
+		// Do not push directories>
 		if f.IsDir() {
 			return nil
 		}
@@ -60,7 +63,9 @@ func TestPutBranch(t *testing.T) {
 		key := strings.TrimPrefix(path, repository.WorkDir(repo))
 		kvPath := filepath.Join(repo.Name(), repo.Branch().Short(), key)
 		kvContent, _, err := handler.Get(kvPath, nil) //nolint:ineffassign,staticcheck
-		fileContent, err := ioutil.ReadFile(path)     //nolint:ineffassign,staticcheck
+		assert.NoError(t, err)
+		fileContent, err := os.ReadFile(path) //nolint:ineffassign,staticcheck
+		assert.NoError(t, err)
 
 		assert.Equal(t, fileContent, kvContent.Value)
 		return nil
