@@ -19,7 +19,6 @@ package kv
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -27,7 +26,7 @@ import (
 
 	"github.com/Aireuropa/git2consul-go/repository"
 	"github.com/stretchr/testify/assert"
-	yaml "gopkg.in/yaml.v2"
+	yaml "gopkg.in/yaml.v3"
 )
 
 type mockHandler struct {
@@ -39,7 +38,7 @@ var (
 	yamlFile File
 	textFile File
 	handler  Handler
-	yamlTree map[interface{}]interface{}
+	yamlTree map[string]interface{}
 	keys     map[string][]byte
 )
 
@@ -56,19 +55,19 @@ const (
 		"    - second_element: value_4\n"
 )
 
-//TestFile performs tests on implemented file handlers.
+// TestFile performs tests on implemented file handlers.
 // * yaml
 // * text
 func TestFileHandler(t *testing.T) {
 	var repo repository.Repo
-	yamlTree = make(map[interface{}]interface{})
+	yamlTree = make(map[string]interface{})
 	err := yaml.Unmarshal([]byte(content), &yamlTree)
 	if err != nil {
 		t.Fatal(err)
 	}
 	filePath := filepath.Join(os.TempDir(), "foo.yml")
 	defer os.Remove(filePath)
-	err = ioutil.WriteFile(filePath, []byte(content), 0700)
+	err = os.WriteFile(filePath, []byte(content), 0700)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -85,7 +84,7 @@ func TestFileHandler(t *testing.T) {
 	t.Run("TestDeleteTextFile", func(t *testing.T) { testDeleteTextFile(t, repo) })
 }
 
-//testParsNodes verfies yaml file evaluation function.
+// testParsNodes verfies yaml file evaluation function.
 func testParseYamlEntries(t *testing.T) {
 	keys := entriesToKV(yamlTree)
 	if string(keys["ei_unix_cavisson::cavisson_collector_srv"]) != "10.206.96.18" {
