@@ -19,8 +19,8 @@ package repository
 import (
 	"fmt"
 
-	"gopkg.in/src-d/go-git.v4"
-	"gopkg.in/src-d/go-git.v4/plumbing"
+	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing"
 )
 
 // Pull a repository branch, which is equivalent to a fetch and merge
