@@ -23,11 +23,11 @@ import (
 	"sync"
 
 	"github.com/Aireuropa/git2consul-go/config"
-	"gopkg.in/src-d/go-git.v4"
-	"gopkg.in/src-d/go-git.v4/plumbing"
-	"gopkg.in/src-d/go-git.v4/plumbing/object"
-	"gopkg.in/src-d/go-git.v4/plumbing/transport"
-	"gopkg.in/src-d/go-git.v4/storage"
+	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing"
+	"github.com/go-git/go-git/v5/plumbing/object"
+	"github.com/go-git/go-git/v5/plumbing/transport"
+	"github.com/go-git/go-git/v5/storage"
 )
 
 // Repo interface represents Repository
@@ -143,7 +143,7 @@ func (r *Repository) init(repoPath string) (int, error) {
 	return RepositoryOpened, nil
 }
 
-//WorkDir returns working directory for a local copy of the repository.
+// WorkDir returns working directory for a local copy of the repository.
 func WorkDir(r Repo) string {
 	w, _ := r.Worktree()
 	return w.Filesystem.Root()
